// src/app/models/member-profile.ts
export class MemberProfile {
  name: string;
  email: string;
  faculty: string;
  program: string;
  programStartDate: string;
  programEndDate: string;
  role: string;
  requestType: string;
  requestDate: string;
  requestStatus: string;
  accessModificationType: string;
  accessModificationDate: string;
  accessStatus: string; //removed or deactivated
  roleModificationDate: string;
  mode: string; //V - View; C - Create
}